+++ apps/meteor/server/services/gso/fake/index.ts
@@ -1,0 +1,2 @@
+export * from './employees';
+export * from './tasks';
